pyproject.toml
[project]
name = "krishisarthi-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "asyncpg==0.29.0",
    "fastapi==0.111.0",
    "httpx==0.27.0",
    "joblib==1.4.2",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pydantic>=2.7.4",
    "python-dotenv==1.0.1",
    "google-auth==2.35.0",
    "scikit-learn>=1.5.2",
    "uvicorn[standard]==0.30.1",
    "xgboost==2.0.3",
    # HTTP helpers
    "requests>=2.31.0",
    # LangGraph / LangChain chatbot stack
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langsmith>=0.1.0",
    # Sync-inside-async helper
    "nest-asyncio>=1.6.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["chatbot", "services"]
include-package-data = true

[dependency-groups]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
]
